pyproject.toml
[project]
name = "githubctl"
dynamic = ["version"]
description = "github command line tool."
authors = [
    {name = "Hamza Y", email = "letters2hamza@gmail.com"},
]
dependencies = [
    "ruff>=0.5.3",
    "typer[all]>=0.12.3",
    "pydantic>=2.8.2",
    "jmespath>=1.0.1",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "pytest>=8.3.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
githubctl = "githubctl.cli:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
version = { source = "file", path = "src/githubctl/__init__.py" }
distribution = true

[tool.pdm.build]
source-includes = [
    "tests/",
    "requirements*.txt",
]

[tool.mypy]
strict = true
exclude = "tests/assets/*"

[tool.ruff.format]
line-ending = "auto"
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191", # indentation contains tabs
]

